% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateCNVtree.R
\name{annotateCNVTree}
\alias{annotateCNVTree}
\title{Annotate a Phylogenetic Tree with CNV Events}
\usage{
annotateCNVTree(tree, cnv_mat, cnv_thresh = 0.15)
}
\arguments{
\item{tree}{A phylogenetic tree (of class \code{phylo}) that will be annotated.}

\item{cnv_mat}{A matrix of copy number variation (CNV) values, with samples as rows and regions as columns.}

\item{cnv_thresh}{A numeric threshold to filter significant CNV events. Default is 0.15.}
}
\value{
A data frame with the tree data, including annotations for CNV events.
}
\description{
This function annotates a phylogenetic tree with copy number variation (CNV) events. It identifies
significant CNV events in the provided matrix, links them to clones and ancestral nodes, and
updates the tree with this information.
}
\examples{
cnv_matrix <- structure(c(0.2, 0.4, 0, 0, 0.1, 0, 0.1, 0.2, 0.2), dim = c(3L,
3L), dimnames = list(c("Clone 1", "Clone 2", "Clone 3"), c("Region 1",
                                                          "Region 2", "Region 3")))
tree <- buildCNVTree(cnv_matrix)
tree_data <- annotateCNVTree(tree, cnv_matrix)

}
