% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCNVTree.R
\name{buildCNVTree}
\alias{buildCNVTree}
\title{Construct a Phylogenetic Tree from a Copy Number Variation (CNV) Matrix}
\usage{
buildCNVTree(cnv_matrix, tree_function = nj, dist_method = "euclidean")
}
\arguments{
\item{cnv_matrix}{A matrix representing copy number variation, where rows correspond to samples
and columns correspond to genomic regions. Each value represents the CNV at a given region in a sample.}

\item{tree_function}{A function to construct the phylogenetic tree from a distance matrix. The default is
\code{nj} (Neighbor-Joining). Other functions (e.g., \code{upgma}, \code{wpgma}) can also be used.}

\item{dist_method}{The distance method to be used.}
}
\value{
A rooted phylogenetic tree (of class \code{phylo})
}
\description{
This function constructs a phylogenetic tree based on a given copy number variation (CNV) matrix.
It adds a baseline "Normal" profile only to root the tree, which is not shown in final output.
First, it computes pairwise distances between profiles using Euclidean distance, and then applies a specified tree-building function
(e.g., Neighbor-Joining) to construct the tree.
}
\examples{
# Example usage with Neighbor-Joining (default)

}
