% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cit.density.R
\name{cit.density}
\alias{cit.density}
\title{cit.density}
\usage{
cit.density(x, doplot = FALSE, pc = 0.05, ...)
}
\arguments{
\item{x}{...}

\item{doplot}{...}

\item{pc}{...}

\item{...}{...}
}
\description{
cit.density
}
