% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cit.discretize.R
\name{cit.discretize}
\alias{cit.discretize}
\title{cit.discretize}
\usage{
cit.discretize(x, lim, quant = FALSE, addlevels = FALSE)
}
\arguments{
\item{x}{...}

\item{lim}{...}

\item{quant}{...}

\item{addlevels}{...}
}
\description{
cit.discretize
}
