% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cit.peaks.R
\name{cit.peaks}
\alias{cit.peaks}
\title{cit.peaks}
\usage{
cit.peaks(
  x,
  percentHighestPeak = 0.2,
  maxNbPeaks = NULL,
  minDeltaBetweenPeaks = 0.03,
  deltaApproach = 1,
  doplot = FALSE,
  bw = "nrd0",
  ...
)
}
\arguments{
\item{x}{...}

\item{percentHighestPeak}{...}

\item{maxNbPeaks}{...}

\item{minDeltaBetweenPeaks}{...}

\item{deltaApproach}{...}

\item{doplot}{...}

\item{bw}{...}

\item{...}{...}
}
\value{
...
}
\description{
cit.peaks
}
