% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVclassify.R
\name{classify_cnv}
\alias{classify_cnv}
\title{CNVclassify}
\usage{
classify_cnv(cnv_vector, peaks)
}
\arguments{
\item{cnv_vector}{Vector of CNVs}

\item{peaks_info}{peaks info}
}
\description{
CNVclassify
}
