% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVCallingList.R
\name{CNVcallingList}
\alias{CNVcallingList}
\title{CNVcalling but for a list of seurat objects}
\usage{
CNVcallingList(
  seuratList,
  assay = NULL,
  referenceVar = NULL,
  referenceLabel = NULL,
  scaleOnReferenceLabel = TRUE,
  denoise = TRUE,
  thresholdPercentile = 0.01,
  geneMetadata = getGenes(),
  windowSize = 150,
  windowStep = 10,
  saveGenomicWindows = FALSE,
  topNGenes = 7000
)
}
\arguments{
\item{seuratList}{List of seurat objects}

\item{assay}{Name of the assay to run the CNV on. Takes the results of \code{prepareCountsForCNVAnalysis} by default if available}

\item{referenceVar}{The variable name of the annotations in the Seurat metadata}

\item{referenceLabel}{The label given to the observations wanted as reference (can be any type of annotation)}

\item{scaleOnReferenceLabel}{If you want to scale the results depending on the reference observations}

\item{denoise}{If the data needs to be denoised (default = \code{TRUE})}

\item{thresholdPercentile}{Which quantiles to take (if 0.01 it will take 0.01-0.99). Background noise appears with higher numbers.}

\item{geneMetadata}{List of genes from ensembl and their metadata}

\item{windowSize}{Size of the genomic windows}

\item{windowStep}{Step between the genomic windows}

\item{saveGenomicWindows}{If the information of the genomic windows need to be saved in the current directory (default = \code{FALSE}).}

\item{topNGenes}{Number of top expressed genes}
}
\value{
This function returns a list of the genomic scores per genomic window
}
\description{
CNVcalling but for a list of seurat objects
}
