% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVclassification.R
\name{CNVclassification}
\alias{CNVclassification}
\title{CNV Classification
Classifies the CNV results into loss, gain, or no alteration for each observation and chromosome arm.}
\usage{
CNVclassification(seuratObj, peaks = c(-0.1, 0, 0.1))
}
\arguments{
\item{seuratObj}{A Seurat object containing the results of the CNV analysis (e.g., from \code{fastCNV}).}

\item{peaks}{A numeric vector containing the thresholds for classifying CNVs. The default is \code{c(-0.1, 0, 0.1)}, which defines:
\itemize{
\item Loss: CNV scores below \code{-0.1}
\item No alteration: CNV scores between \code{-0.1} and \code{0.1}
\item Gain: CNV scores above \code{0.1}
}}
}
\value{
The same Seurat object with an additional classification for each observation and chromosome arm in the metadata.
The classification can be one of \code{"loss"}, \code{"gain"}, or \code{"no_alteration"}.
}
\description{
CNV Classification
Classifies the CNV results into loss, gain, or no alteration for each observation and chromosome arm.
}
