% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVcluster.R
\name{CNVCluster}
\alias{CNVCluster}
\title{Perform CNV Clustering with Seurat Object}
\usage{
CNVCluster(seuratObj, k = NULL, h = NULL)
}
\arguments{
\item{seuratObj}{A Seurat object containing a "genomicScores" assay with a matrix of genomic scores for clustering.}

\item{k}{Optional. The number of clusters to cut the dendrogram into. If \code{NULL}, the optimal number of clusters is determined automatically using the elbow method.}

\item{h}{Optional. The height at which to cut the dendrogram for clustering. If both \code{k} and \code{h} are provided, \code{k} takes precedence.}
}
\value{
A Seurat object with an additional metadata column, \code{cnv_clusters}, containing the cluster assignments.
}
\description{
The \code{CNVcluster} function performs hierarchical clustering on a genomic score matrix extracted from a Seurat object.
It provides options for plotting a dendrogram, an elbow plot for optimal cluster determination,
and cluster visualization on the dendrogram. The resulting cluster assignments are stored in the Seurat object.
}
\details{
The function computes a Manhattan distance matrix and performs hierarchical clustering using the Ward.D2 method.
If \code{k} is not provided, the elbow method is applied to determine the optimal number of clusters based on the within-cluster sum of squares (WSS).

The clusters are assigned to the Seurat object under the metadata column \code{cnv_clusters}.
}
