% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVfractionPerChromosome.R
\name{CNVfractionPerChromosome}
\alias{CNVfractionPerChromosome}
\title{CNVfractionPerChromosome
Computes the CNV fraction of each spot/cell per chromosome arm, then stocks the results into the metadata}
\usage{
CNVfractionPerChromosome(seuratObject)
}
\arguments{
\item{seuratObject}{The output of fastCNV()}
}
\value{
This function returns the same seurat object with the CNV fraction for each chromosome arm in the metadata
}
\description{
CNVfractionPerChromosome
Computes the CNV fraction of each spot/cell per chromosome arm, then stocks the results into the metadata
}
