% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVPerChromosomeArm.R
\name{CNVPerChromosomeArm}
\alias{CNVPerChromosomeArm}
\title{CNVPerChromosomeArm
Computes the CNV fraction of each spot/cell per chromosome arm, then stocks the results into the metadata}
\usage{
CNVPerChromosomeArm(seuratObject)
}
\arguments{
\item{seuratObject}{The output of fastCNV()}
}
\value{
This function returns the same seurat object with the CNV for each chromosome arm in the metadata
}
\description{
CNVPerChromosomeArm
Computes the CNV fraction of each spot/cell per chromosome arm, then stocks the results into the metadata
}
