% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVTree.R
\name{CNVTree}
\alias{CNVTree}
\title{Build, annotate and plot a Phylogenetic Tree from a seurat Object containing the CNV results from \code{fastCNV()}}
\usage{
CNVTree(
  seuratObj,
  healthyClusters = NULL,
  values = "scores",
  cnv_thresh = 0.15,
  tree_function = nj,
  dist_method = "euclidean",
  clone_cols = TRUE
)
}
\arguments{
\item{seuratObj}{A Seurat object containing CNV data and metadata.}

\item{healthyClusters}{A numeric vector or \code{NULL}. If provided, clusters specified in this vector
will be labeled as "Benign" instead of "Clone". Default is \code{NULL}.}

\item{values}{one of 'scores' or 'calls'. 'scores' returns the mean CNV score per cluster,
while 'calls' uses \code{cnv_thresh} to establish a cut-off for gains and losses, returning a matrix
of CNV calls (0=none, 1=gain, -1=loss).}

\item{cnv_thresh}{A numeric threshold to filter significant CNV events. Default is 0.15.}

\item{tree_function}{A function to construct the phylogenetic tree from a distance matrix. The default is
\code{nj} (Neighbor-Joining). Other functions (e.g., \code{upgma}, \code{wpgma}) can also be used.}

\item{dist_method}{The distance method to be used.}

\item{clone_cols}{a color palette to color the clones. If NULL, points are
not colored. If TRUE, clones are colored using default color palette. If a
palette is given, clones are colored following the palette, with
values passed to \code{scale_color_manual}. Custom palettes should contain 
at least as many colors as the number of clones in the tree.}
}
\description{
Build, annotate and plot a Phylogenetic Tree from a seurat Object containing the CNV results from \code{fastCNV()}
}
