% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAverageExpression.R
\name{computeAverageExpression}
\alias{computeAverageExpression}
\title{Compute average expression for patients}
\usage{
computeAverageExpression(LN, LrawcountsByPatient)
}
\arguments{
\item{LN}{A list where each element represents a cell type with sublists containing patient data.}

\item{LrawcountsByPatient}{A named list where each element contains count data for a specific patient.}
}
\value{
A named vector containing the average expression for each patient.
}
\description{
This function calculates the average gene expression for each patient across
different cell types. It first retrieves patient data from LN, then extracts
the corresponding count data from LrawcountsByPatient, and calculates the mean expression.
}
