% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.fastCNV.R
\docType{package}
\name{fastCNV-package}
\alias{fastCNV-package}
\title{fastCNV: fastCNV}
\description{
This package performs fast CNV analysis on Seurat data, providing tools for detecting and visualizing
Copy Number Variations (CNVs) at both individual and group levels. It includes various functions
for CNV detection, clustering, visualization, and data handling, specifically designed for single-cell
and spatial transcriptomics datasets processed using Seurat.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://must-bioinfo.github.io/fastCNV}
}

}
\author{
\strong{Maintainer}: Gadea Cabrejas \email{gadea.cabrejas-saiz@u-paris.fr}

Authors:
\itemize{
  \item Clarice Groeneveld \email{clarice.groeneveld@inserm.fr} [contributor]
}

}
\keyword{internal}
