% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastCNV.R
\name{fastCNV}
\alias{fastCNV}
\title{fastCNV calls all of the internal functions needed to compute the putative CNV on a seurat object or a list of seurat objects}
\usage{
fastCNV(
  seuratObj,
  sampleName,
  referenceVar = NULL,
  referenceLabel = NULL,
  splitPlotOnVar = referenceVar,
  downsizePlot = FALSE,
  doPlot = TRUE,
  printPlot = FALSE,
  prepareCounts = TRUE,
  pooledReference = TRUE,
  assay = NULL,
  getCNVPerChromosomeArm = TRUE,
  savePath = ".",
  denoise = TRUE,
  aggregFactor = 15000,
  seuratClusterResolution = 0.8,
  aggregateByVar = T,
  reClusterSeurat = F,
  scaleOnReferenceLabel = TRUE,
  thresholdPercentile = 0.01,
  geneMetadata = getGenes(),
  windowSize = 150,
  windowStep = 10,
  saveGenomicWindows = FALSE,
  topNGenes = 7000
)
}
\arguments{
\item{seuratObj}{Seurat object or list of seurat objects}

\item{sampleName}{Name of the sample or list of names corresponding to the samples in the \code{seuratObj}}

\item{referenceVar}{The variable name of the annotations in the Seurat metadata}

\item{referenceLabel}{The label given to the observations wanted as reference (can be any type of annotation)}

\item{splitPlotOnVar}{The name given to the meta.data slot to split the observations on during the \code{plotCNVResults} step, if different from \code{referenceVar}.}

\item{downsizePlot}{Subset the observations to speed up the plotting process (default = \code{FALSE}).}

\item{doPlot}{If \code{TRUE} will build a heatmap for each of the samples (default = \code{TRUE})}

\item{printPlot}{If the heatmap should be printed in the console (default = \code{FALSE}, the plot will only be in a PDF).}

\item{prepareCounts}{If \code{FALSE} will not run the \code{prepareCountsForCNVAnalysis} function. Default \code{TRUE}.}

\item{pooledReference}{Default \code{TRUE}. Will build a pooled reference across all samples if \code{TRUE}.}

\item{assay}{Name of the assay to run the CNV on. Takes the results of \code{prepareCountsForCNVAnalysis} by default if available}

\item{getCNVPerChromosomeArm}{If \code{TRUE}, will save into the metadata the CNV per chromosome arm}

\item{savePath}{Default \code{.}. If \code{NULL} the heatmap won't be saved as a pdf.}

\item{denoise}{If the data needs to be denoised (default = \code{TRUE})}

\item{aggregFactor}{The number of counts per spot desired (default \verb{30 000}). If less than \code{1000}, will not run the \code{prepareCountsForCNVAnalysis} function.}

\item{seuratClusterResolution}{The resolution wanted for the seurat clusters (default 0.8)}

\item{aggregateByVar}{If \code{referenceVar} is given, whether to use it to pool the observations}

\item{reClusterSeurat}{Whether to re-cluster if the Seurat object given already has a \code{seurat_clusters} slot in its meta.data}

\item{scaleOnReferenceLabel}{If you want to scale the results depending on the normal observations}

\item{thresholdPercentile}{Which quantiles to take (if 0.01 it will take 0.01-0.99). Background noise appears with higher numbers.}

\item{geneMetadata}{List of genes and their metadata (default genes from ensembl version 113)}

\item{windowSize}{Size of the genomic windows}

\item{windowStep}{Step between the genomic windows}

\item{saveGenomicWindows}{If the information of the genomic windows need to be saved in the current directory (default = \code{FALSE}).}

\item{topNGenes}{Number of top expressed genes to keep}
}
\value{
This function returns a list of the seurat objects after all the analysis and it creates the heatmaps of the CNVs of every object in seuratObj and saves them in a pdf in the current working directory
}
\description{
fastCNV calls all of the internal functions needed to compute the putative CNV on a seurat object or a list of seurat objects
}
