% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.fastCNV.R
\docType{package}
\name{fastCNV2-package}
\alias{fastCNV2}
\alias{fastCNV2-package}
\title{fastCNV2: fastCNV2 - Fast CNV Detection with Enhanced Compatibility (Modified Version)}
\description{
This package performs fast CNV analysis on Seurat data, providing tools for detecting and visualizing
Copy Number Variations (CNVs) at both individual and group levels. It includes various functions
for CNV detection, clustering, visualization, and data handling, specifically designed for single-cell
and spatial transcriptomics datasets processed using Seurat.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Zaoqu-Liu/fastCNV2}
  \item Report bugs at \url{https://github.com/Zaoqu-Liu/fastCNV2/issues}
}

}
\author{
\strong{Maintainer}: Zaoqu Liu \email{liuzaoqu@163.com} (Modified for enhanced compatibility) [contributor]

Authors:
\itemize{
  \item Gadea Cabrejas \email{gadea.cabrejas-saiz@u-paris.fr} (Original author) [copyright holder]
  \item Clarice Groeneveld \email{clarice.groeneveld@inserm.fr} (Original author)
}

}
\keyword{internal}
