% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastCNV_10XHD.R
\name{fastCNV_10XHD}
\alias{fastCNV_10XHD}
\title{fastCNV_10XHD calls all of the internal functions needed to compute the putative CNV on a Seurat Visium HD object or a list of Seurat Visium HD objects}
\usage{
fastCNV_10XHD(
  seuratObjHD,
  sampleName,
  referenceVar = NULL,
  referenceLabel = NULL,
  assay = "Spatial.016um",
  pooledReference = TRUE,
  scaleOnReferenceLabel = TRUE,
  thresholdPercentile = 0.01,
  geneMetadata = getGenes(),
  windowSize = 150,
  windowStep = 10,
  saveGenomicWindows = FALSE,
  topNGenes = 7000,
  chrArmsToForce = NULL,
  genesToForce = NULL,
  regionToForce = NULL,
  getCNVPerChromosomeArm = TRUE,
  getCNVClusters = TRUE,
  k_clusters = NULL,
  h_clusters = NULL,
  cellTypesToCluster = NULL,
  mergeCNV = TRUE,
  mergeThreshold = 0.98,
  doPlot = TRUE,
  denoise = TRUE,
  printPlot = FALSE,
  savePath = ".",
  outputType = "png",
  clustersVar = "cnv_clusters",
  clusters_palette = "default",
  splitPlotOnVar = clustersVar,
  referencePalette = "default",
  raster_resize_mat = TRUE
)
}
\arguments{
\item{seuratObjHD}{Seurat object or list of Seurat objects to perform the CNV analysis on.}

\item{sampleName}{Name of the sample or a list of names corresponding to the samples in the \code{seuratObj}.}

\item{referenceVar}{The variable name of the annotations in the Seurat metadata to be used as reference.}

\item{referenceLabel}{The label given to the observations you want as reference (can be any type of annotation).}

\item{assay}{Name of the assay to run the CNV on. Takes the results of \code{prepareCountsForCNVAnalysis} by default if available.}

\item{pooledReference}{Default is \code{TRUE}. Will build a pooled reference across all samples if \code{TRUE}.}

\item{scaleOnReferenceLabel}{If \code{TRUE}, scales the results depending on the normal observations (default = \code{TRUE}).}

\item{thresholdPercentile}{Which quantiles to take (default 0.01). For example, \code{0.01} will take quantiles between 0.01-0.99. Background noise appears with higher numbers.}

\item{geneMetadata}{List of genes and their metadata (default uses genes from Ensembl version 113).}

\item{windowSize}{Size of the genomic windows for CNV analysis (default = 150).}

\item{windowStep}{Step between the genomic windows (default = 10).}

\item{saveGenomicWindows}{If \code{TRUE}, saves the information of the genomic windows in the current directory (default = \code{FALSE}).}

\item{topNGenes}{Number of top expressed genes to keep (default = 7000).}

\item{chrArmsToForce}{A chromosome arm (e.g., \code{"8p"}, \code{"3q"}) or a list of chromosome arms (e.g., \code{c("3q", "8p", "17p")}) to force into the analysis.}

\item{genesToForce}{A list of genes to force into the analysis (e.g. \code{c("FOXP3","MUC16","SAMD15")}).}

\item{regionToForce}{Chromosome region to force into the analysis (vector containing chr, start, end).}

\item{getCNVPerChromosomeArm}{If \code{TRUE}, will save the CNV per chromosome arm into the metadata.}

\item{getCNVClusters}{If \code{TRUE}, will perform clustering on the CNV scores and save them in the metadata of the Seurat object as \code{cnv_clusters}.}

\item{k_clusters}{Optional. Number of clusters to cut the dendrogram into. If \code{NULL}, the optimal number of clusters is determined automatically using the elbow method.}

\item{h_clusters}{Optional. The height at which to cut the dendrogram for clustering. If both \code{k} and \code{h} are provided, \code{k} takes precedence.}

\item{cellTypesToCluster}{Optional. Cell type annotations to use for clustering. If \code{NULL} (default), all observations will be clustered.}

\item{mergeCNV}{Logical. Whether to merge the highly correlated CNV clusters.}

\item{mergeThreshold}{A numeric value between 0 and 1. Clusters with correlation greater than this threshold will be merged. Default is 0.98.}

\item{doPlot}{If \code{TRUE}, will build a heatmap for each of the samples (default = \code{TRUE}).}

\item{denoise}{If \code{TRUE}, the denoised data will be used in the heatmap (default = \code{TRUE}).}

\item{printPlot}{If \code{TRUE}, the heatmap will be printed in the console (default = \code{FALSE}, the plot will only be saved in a PDF).}

\item{savePath}{Path to save the heatmap plot. If \code{NULL}, the plot won't be saved (default = \code{.}).}

\item{outputType}{Specifies the file format for saving the plot, either \code{"png"} or \code{"pdf"} (default = \code{"png"}).}

\item{clustersVar}{The name of the metadata column containing cluster information (default = \code{"cnv_clusters"}).}

\item{clusters_palette}{A color palette for \code{clustersVar}.
You can provide a custom palette as a vector of color codes (e.g., \code{c("#F8766D", "#A3A500", "#00BF7D")}).}

\item{splitPlotOnVar}{The name of the metadata column to split the observations during the \code{plotCNVResults} step, if different from \code{referenceVar}.}

\item{referencePalette}{A color palette for \code{referenceVar}.
You can provide a custom palette as a vector of color codes (e.g., \code{c("#FF0000", "#00FF00")}).}

\item{raster_resize_mat}{Whether resize the matrix to let the dimension of the matrix the same as the dimension of the raster image.
Default is TRUE.}
}
\value{
A Seurat object or a list of Seurat objects after all the analysis is complete. Heatmaps of the CNVs for every object in \code{seuratObj} are generated and saved in the specified path (default = current working directory).
}
\description{
This function orchestrates the CNV analysis on a Seurat Visium HD object (or multiple objects). It calls internal functions such as
\code{CNVAnalysis} and \code{PlotCNVResults} to compute the CNVs and generate heatmaps. The results are saved in the metadata of the Seurat object(s), with options for
generating and saving plots.
}
