% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenes.R
\name{getGenes}
\alias{getGenes}
\title{Downloads the gene information from the ensembl site}
\usage{
getGenes(filters = NULL, cache = TRUE)
}
\arguments{
\item{filters}{Filters (one or more) that should be used in the query}

\item{cache}{Whether to use the data from the cache or to download the latest version}
}
\value{
The function returns a list of all the genes and the information related to these genes
}
\description{
Downloads the gene information from the ensembl site
}
