% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ubuntu22.R
\name{is_ubuntu22}
\alias{is_ubuntu22}
\title{Checks if working on ubuntu 22}
\usage{
is_ubuntu22()
}
\value{
TRUE or FALSE
}
\description{
This function checks if the user is working on ubuntu 22
Raster_by_magick does not work under ubuntu 22 so the parameter on the heatmap is changed if it returns TRUE
}
