% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeCNVClusters.R
\name{mergeCNVClusters}
\alias{mergeCNVClusters}
\title{Merge CNV Clusters in a Seurat Object}
\usage{
mergeCNVClusters(seuratObj, mergeThreshold = 0.98)
}
\arguments{
\item{seuratObj}{A Seurat object containing fastCNV's results by chromosome arm, and CNV clustering.}

\item{mergeThreshold}{A numeric value between 0 and 1. Clusters with correlation
greater than this threshold will be merged. Default is 0.98.}
}
\value{
A Seurat Object with updated CNV clusters, where highly correlated clusters have been merged.
}
\description{
This function merges CNV clusters in a Seurat object based on the correlation
of their average CNV profiles across chromosome arms. Clusters with correlation
greater than a user-specified threshold are merged into a single cluster.
}
