% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNVResults.R
\name{plotCNVResults}
\alias{plotCNVResults}
\title{A function to plot the CNV results into a heatmap}
\usage{
plotCNVResults(
  seuratObj,
  referenceVar = NULL,
  clustersVar = "cnv_clusters",
  splitPlotOnVar = clustersVar,
  savePath = ".",
  printPlot = FALSE,
  referencePalette = as.character(paletteer::paletteer_d("pals::glasbey")),
  clusters_palette = scales::hue_pal()
)
}
\arguments{
\item{seuratObj}{Seurat object containing the data used to get the genomicScores.}

\item{referenceVar}{The variable name of the annotations in the Seurat object metadata}

\item{clustersVar}{The variable name of the clusters in the Seurat object metadata}

\item{splitPlotOnVar}{The variable name on which to split the heatmap rows.}

\item{savePath}{Path to save the pdf heatmap. If \code{NULL}, plot won't be saved (default = \code{.}).}

\item{printPlot}{If the heatmap should be printed in the console.}

\item{referencePalette}{The color palette that should be used for \code{referenceVar}}

\item{clusters_palette}{The color palette that should be used for \code{clustersVar}}
}
\value{
This function builds a heatmap and saves it as a .pdf file in your working directory
}
\description{
A function to plot the CNV results into a heatmap
}
