% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNVResultsHD.R
\name{plotCNVResultsHD}
\alias{plotCNVResultsHD}
\title{Plot Visium HD CNV Results into a Heatmap
Builds a heatmap to visualize the Visium HD CNV results based on genomic scores.}
\usage{
plotCNVResultsHD(
  seuratObjHD,
  referenceVar = NULL,
  splitPlotOnVar = referenceVar,
  savePath = ".",
  printPlot = FALSE,
  referencePalette = "default",
  outputType = "png"
)
}
\arguments{
\item{seuratObjHD}{A Seurat object containing the genomic scores computed previously.}

\item{referenceVar}{The name of the metadata column in the Seurat object containing reference annotations.}

\item{splitPlotOnVar}{The name of the metadata column used to split the heatmap rows (e.g., cell type or cluster) (default = \code{clustersVar}).}

\item{savePath}{The path where the heatmap will be saved. If \code{NULL}, the plot will not be saved (default = \code{"."}).}

\item{printPlot}{Logical. If \code{TRUE}, prints the heatmap to the console.}

\item{referencePalette}{A color palette for \code{referenceVar}.
You can provide a custom palette as a vector of color codes (e.g., \code{c("#FF0000", "#00FF00")}).}

\item{outputType}{Character. Specifies the file format for saving the plot, either \code{"png"} or \code{"pdf"}.}
}
\value{
This function generates a heatmap and saves it as a \code{.pdf} or \code{.png} file in the specified path (default = working directory).
}
\description{
Plot Visium HD CNV Results into a Heatmap
Builds a heatmap to visualize the Visium HD CNV results based on genomic scores.
}
