% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVanalysis.R
\name{CNVanalysis}
\alias{CNVanalysis}
\title{CNVanalysis
Runs Copy Number Variation (CNV) analysis on a Seurat object or a list of Seurat objects.}
\usage{
CNVanalysis(
  object,
  referenceVar = NULL,
  referenceLabel = NULL,
  pooledReference = TRUE,
  scaleOnReferenceLabel = TRUE,
  denoise = TRUE,
  assay = NULL,
  thresholdPercentile = 0.01,
  geneMetadata = getGenes(),
  windowSize = 150,
  windowStep = 10,
  saveGenomicWindows = FALSE,
  topNGenes = 7000,
  chrArmsToForce = NULL,
  genesToForce = NULL,
  regionToForce = NULL
)
}
\arguments{
\item{object}{A Seurat object or a list of Seurat objects containing the data for CNV analysis. Each object
can be either \strong{single-cell} or \strong{spatial transcriptomics} data.}

\item{referenceVar}{The name of the metadata column in the Seurat object that contains reference annotations.}

\item{referenceLabel}{The label within \code{referenceVar} that specifies the reference population (can be any type of annotation).}

\item{pooledReference}{Logical. If \code{TRUE} (default), builds a pooled reference across all samples.}

\item{scaleOnReferenceLabel}{Logical. If \code{TRUE} (default), scales the results based on the reference population.}

\item{denoise}{Logical. If \code{TRUE} (default), applies denoising to the data.}

\item{assay}{Name of the assay to run the CNV analysis on. Defaults to the results of \code{prepareCountsForCNVAnalysis} if available.}

\item{thresholdPercentile}{Numeric. Specifies the quantile range to consider (e.g., \code{0.01} keeps values between the 1st and 99th percentiles). Higher values filter out more background noise.}

\item{geneMetadata}{A dataframe containing gene metadata, typically from Ensembl.}

\item{windowSize}{Integer. Defines the size of genomic windows for CNV analysis.}

\item{windowStep}{Integer. Specifies the step size between genomic windows.}

\item{saveGenomicWindows}{Logical. If \code{TRUE}, saves genomic window information in the current directory (default = \code{FALSE}).}

\item{topNGenes}{Integer. The number of top-expressed genes to retain in the analysis.}

\item{chrArmsToForce}{A chromosome arm (e.g., \code{"8p"}, \code{"3q"}) or a list of chromosome arms (e.g., \code{c("3q", "8p", "17p")}) to force into the analysis.
If specified, all genes within the given chromosome arm(s) will be included.}

\item{genesToForce}{A list of genes to force into the analysis (e.g. \code{c("FOXP3","MUC16","SAMD15")}).}

\item{regionToForce}{Chromosome region to force into the analysis (vector containing chr, start, end).}
}
\value{
If given a \strong{single} Seurat object, returns the same object with:
\itemize{
\item An \strong{additional assay} containing genomic scores per genomic window.
\item A new \strong{CNV fraction column} added to the object’s metadata.
If given a \strong{list} of Seurat objects, returns the modified list.
}
}
\description{
This function performs CNV analysis by calculating genomic scores, applying optional denoising, and optionally
scaling the results based on a reference population. It processes single-cell or spatial transcriptomics data,
generating an additional assay with genomic scores and adding a new metadata column for CNV fractions.
}
