% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVcalling.R
\name{CNVcalling}
\alias{CNVcalling}
\title{CNVcalling
Runs a CNV analysis given a seurat object}
\usage{
CNVcalling(
  seuratObj,
  assay = NULL,
  referenceVar = NULL,
  referenceLabel = NULL,
  scaleOnReferenceLabel = TRUE,
  thresholdPercentile = 0.01,
  genes = getGenes(),
  windowSize = 100,
  windowStep = 20,
  topNGenes = 7000
)
}
\arguments{
\item{seuratObj}{The Seurat object containing the data to analyze}

\item{assay}{Name of the assay to run the CNV on. Takes the results of \code{prepareCountsForCNVAnalysis} by default if available}

\item{referenceVar}{The variable name of the annotations in the Seurat metadata}

\item{referenceLabel}{The label given to the observations wanted as reference (can be any type of annotation)}

\item{scaleOnReferenceLabel}{If you want to scale the results depending on the reference observations}

\item{thresholdPercentile}{which quantiles to take (if 0.01 it will take 0.01-0.99). Background noise appears with higher numbers.}

\item{genes}{List of genes from ensembl}

\item{windowSize}{Size of the genomic windows}

\item{windowStep}{Step between the genomic windows}

\item{topNGenes}{Number of top expressed genes to keep}
}
\value{
This function returns the genomic scores per genomic window
}
\description{
CNVcalling
Runs a CNV analysis given a seurat object
}
