% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVclassification.R
\name{CNVclassification}
\alias{CNVclassification}
\title{CNVclassification}
\usage{
CNVclassification(seuratObj, tumorLabels)
}
\arguments{
\item{seuratObj}{A seurat object containing the results of fastCNV}

\item{tumorLabels}{A list containing the tumor labels}
}
\value{
A seurat object containing the classification (loss gain no_alteration) for
each observation and each chromosome arm.
}
\description{
CNVclassification
}
