% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVclassification.R
\name{CNVclassification}
\alias{CNVclassification}
\title{CNVclassification}
\usage{
CNVclassification(seuratObj, peaks = c(-0.1, 0, 0.1))
}
\arguments{
\item{seuratObj}{A seurat object containing the results of fastCNV}

\item{peaks}{A list containing the thresholds for loss or gain.}
}
\value{
A seurat object containing the classification (loss gain no_alteration) for
each observation and each chromosome arm.
}
\description{
CNVclassification
}
