% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVcluster.R
\name{CNVcluster}
\alias{CNVcluster}
\title{CNV Clustering within a Seurat Object}
\usage{
CNVcluster(seuratObj, nPCs = 10, resolution = 0.8, k.param = 20)
}
\arguments{
\item{seuratObj}{A Seurat object containing single-cell genomic data.}

\item{nPCs}{The number of principal components to use for the clustering. Default is 10.}

\item{resolution}{Resolution parameter for the clustering algorithm. Affects the granularity of the resulting clusters. Default is 0.8.}

\item{k.param}{Number of nearest neighbors to use in the construction of the SNN graph. Default is 20.}
}
\value{
A Seurat object with CNV-based clusters added
}
\description{
Performs clustering based on Copy Number Variations (CNVs) calculated by \code{CNVcalling}.
}
