% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVPerChromosomeArm.R
\name{CNVPerChromosomeArm}
\alias{CNVPerChromosomeArm}
\title{CNV Per Chromosome Arm
Computes the CNV fraction of each spot/cell per chromosome arm, then stores the results into the metadata.}
\usage{
CNVPerChromosomeArm(seuratObj)
}
\arguments{
\item{seuratObj}{A Seurat object, typically the output from the \code{fastCNV()} function, containing genomic scores for CNV analysis.}
}
\value{
The function returns the same Seurat object with the CNV fraction for each chromosome arm added to the metadata.
}
\description{
CNV Per Chromosome Arm
Computes the CNV fraction of each spot/cell per chromosome arm, then stores the results into the metadata.
}
