% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.fastCNV.R
\docType{package}
\name{fastCNV-package}
\alias{fastCNV-package}
\title{fastCNV: fastCNV}
\description{
This package does a fast CNV on seurat data
}
\author{
\strong{Maintainer}: Gadea Cabrejas \email{gadea.cabrejas-saiz@u-paris.fr}

Authors:
\itemize{
  \item Clarice Groeneveld \email{clarice.groeneveld@inserm.fr} [contributor]
}

}
\keyword{internal}
