% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneMetadata.R
\docType{data}
\name{geneMetadata}
\alias{geneMetadata}
\title{Genes Data from Ensembl Version 113}
\format{
An object of class \code{list}, containing gene information as described above.
}
\source{
Ensembl Genome Browser, Version 113: https://www.ensembl.org/index.html
}
\usage{
data(geneMetadata)
}
\description{
Data downloaded from the Ensembl website (version 113), containing detailed gene information for approximately 76,000 genes.
The dataset includes Ensembl gene IDs, HUGO nomenclature (HGNC symbol), Entrez gene IDs, chromosome locations,
gene biotype, and gene length for each gene.
}
\examples{
data(geneMetadata)
hgnc <- geneMetadata$hgnc_symbol
entrez <- geneMetadata$entrezgene_id
}
\keyword{datasets}
