% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGenes.R
\name{getGenes}
\alias{getGenes}
\title{Download Gene Information from Ensembl}
\usage{
getGenes(filters = NULL, cache = TRUE)
}
\arguments{
\item{filters}{A character vector of filters to be applied in the query. These filters determine
which genes and their associated information are returned from the Ensembl database.}

\item{cache}{Logical. If \code{TRUE}, the function will use cached data if available. If \code{FALSE}, it will
download the latest version of the gene data from Ensembl.}
}
\value{
A list containing gene information retrieved from Ensembl, with each element representing
data for a specific gene (e.g., gene IDs, descriptions, associated attributes).
}
\description{
This function retrieves gene information from the Ensembl database using the specified filters.
It can either fetch the latest data or use cached data if available.
}
