% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNVResults.R
\name{plotCNVResults}
\alias{plotCNVResults}
\title{Plot CNV Results into a Heatmap
Builds a heatmap to visualize the CNV results based on genomic scores.}
\usage{
plotCNVResults(
  seuratObj,
  referenceVar = NULL,
  clustersVar = "cnv_clusters",
  splitPlotOnVar = clustersVar,
  denoise = TRUE,
  savePath = ".",
  printPlot = FALSE,
  referencePalette = "default",
  clusters_palette = "default",
  outputType = "png"
)
}
\arguments{
\item{seuratObj}{A Seurat object containing the genomic scores computed previously.}

\item{referenceVar}{The name of the metadata column in the Seurat object containing reference annotations.}

\item{clustersVar}{The name of the metadata column containing cluster information (default = \code{"cnv_clusters"}).}

\item{splitPlotOnVar}{The name of the metadata column used to split the heatmap rows (e.g., cell type or cluster) (default = \code{clustersVar}).}

\item{denoise}{If \code{TRUE}, the denoised data will be used in the heatmap (default = \code{TRUE}).}

\item{savePath}{The path where the heatmap will be saved. If \code{NULL}, the plot will not be saved (default = \code{"."}).}

\item{printPlot}{Logical. If \code{TRUE}, prints the heatmap to the console.}

\item{referencePalette}{A color palette for \code{referenceVar}.
You can provide a custom palette as a vector of color codes (e.g., \code{c("#FF0000", "#00FF00")}).}

\item{clusters_palette}{A color palette for \code{clustersVar}.
You can provide a custom palette as a vector of color codes (e.g., \code{c("#F8766D", "#A3A500", "#00BF7D")}).}

\item{outputType}{Character. Specifies the file format for saving the plot, either \code{"png"} or \code{"pdf"}.}
}
\value{
This function generates a heatmap and saves it as a \code{.pdf} or \code{.png} file in the specified path (default = working directory).
}
\description{
Plot CNV Results into a Heatmap
Builds a heatmap to visualize the CNV results based on genomic scores.
}
