% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCNVTree.R
\name{plotCNVTree}
\alias{plotCNVTree}
\title{Plot an Annotated Phylogenetic Tree with CNV Events}
\usage{
plotCNVTree(tree_data, clone_cols = NULL)
}
\arguments{
\item{tree_data}{A data frame containing tree structure and annotations,
typically produced by \code{annotateCNVtree}.}

\item{clone_cols}{a color palette to color the clones. If NULL, points are
not colored. If TRUE, clones are colored using default color palette. If a
palette is given, clones are colored following the palette, with
values passed to \code{scale_color_manual}. Custom palettes should contain 
at least as many colors as the number of clones in the tree.}
}
\value{
A \code{ggplot} object representing the annotated phylogenetic tree.
}
\description{
This function generates a plot of an annotated phylogenetic tree using \code{ggtree}.
It displays tip labels, tip points, and labels for CNV events associated with
each node.
}
\examples{
cnv_matrix <- structure(c(0.2, 0.4, 0, 0, 0.1, 0, 0.1, 0.2, 0.2), dim = c(3L,
3L), dimnames = list(c("Clone 1", "Clone 2", "Clone 3"), c("Region 1",
                                                          "Region 2", "Region 3")))
tree <- buildCNVTree(cnv_matrix)
tree_data <- annotateCNVTree(tree, cnv_matrix)
plotCNVTree(tree_data)

}
