% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareCountsForCNV.R
\name{prepareCountsForCNVAnalysis}
\alias{prepareCountsForCNVAnalysis}
\title{A function to aggregate observations with the same cell types to get higher counts per observation.
It will help to have higher counts to compute the CNV}
\usage{
prepareCountsForCNVAnalysis(
  seuratObj,
  sampleName = NULL,
  referenceVar = NULL,
  aggregateByVar = T,
  aggregFactor = 30000,
  seuratClusterResolution = 0.8,
  reClusterSeurat = F
)
}
\arguments{
\item{seuratObj}{Seurat object containing the data}

\item{sampleName}{Name of the sample}

\item{referenceVar}{The variable name of the annotations in the Seurat metadata}

\item{aggregateByVar}{If a referenceVar is given, whether to use it to aggregate the observations depending on their type. Default \code{TRUE}.}

\item{aggregFactor}{The number of counts per observation desired (default 30 000)}

\item{seuratClusterResolution}{The resolution wanted for the seurat clusters (default 0.8)}

\item{reClusterSeurat}{If \code{TRUE}, clustering is re-done for the Seurat object.}
}
\value{
This function returns a Seurat object with the modified count matrix in a new assay called \code{AgregatedCounts} and the Seurat clusters in the metadata
}
\description{
A function to aggregate observations with the same cell types to get higher counts per observation.
It will help to have higher counts to compute the CNV
}
