% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareCountsForCNV.R
\name{prepareCountsForCNVAnalysis}
\alias{prepareCountsForCNVAnalysis}
\title{Aggregate Observations by Cell Type for CNV Analysis
Aggregates observations with the same cell types to increase counts per observation,
improving Copy Number Variation (CNV) computation.}
\usage{
prepareCountsForCNVAnalysis(
  seuratObj,
  sampleName = NULL,
  referenceVar = NULL,
  aggregateByVar = T,
  aggregFactor = 15000,
  clusterResolution = 0.8,
  reCluster = F
)
}
\arguments{
\item{seuratObj}{A Seurat object containing the data.}

\item{sampleName}{A character string specifying the sample name.}

\item{referenceVar}{The name of the metadata column in the Seurat object that contains reference annotations.}

\item{aggregateByVar}{Logical. If \code{TRUE} (default), aggregates observations based on \code{referenceVar} annotations.}

\item{aggregFactor}{Integer. The target number of counts per observation (default = \verb{15 000}).}

\item{clusterResolution}{Numeric. The resolution used for Seurat clustering (default = \code{0.8}).}

\item{reCluster}{Logical. If \code{TRUE}, re-runs clustering on the Seurat object.}
}
\value{
A Seurat object with:
\itemize{
\item A new assay called \strong{"AggregatedCounts"} containing the modified count matrix.
\item \strong{Seurat clusters} stored in the metadata.
}
}
\description{
Aggregate Observations by Cell Type for CNV Analysis
Aggregates observations with the same cell types to increase counts per observation,
improving Copy Number Variation (CNV) computation.
}
